Extracted to a trainers package
- Cleans up some stuff
