+++ client/android/Crowd-ML/app/src/osu/crowd_ml/BackgroundDataSend.java
@@ -32,6 +32,8 @@ import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 import android.util.Log;
 
+import osu.crowd_ml.trainers.TensorFlowTrainer;
+import osu.crowd_ml.trainers.Trainer;
 import osu.crowd_ml.utils.DataSender;
