+++ client/android/Crowd-ML/app/src/osu/crowd_ml/trainers/InternalTrainer.java
@@ -1,4 +1,4 @@
-package osu.crowd_ml;
+package osu.crowd_ml.trainers;
 
 import android.util.Log;
 
@@ -7,6 +7,9 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Random;
 
+import osu.crowd_ml.BuildConfig;
+import osu.crowd_ml.Parameters;
+import osu.crowd_ml.TrainingDataIO;
 import osu.crowd_ml.loss_functions.LossFunction;
 import osu.crowd_ml.noise_distributions.Distribution;
 import osu.crowd_ml.utils.ArrayUtils;
